Don't set prunesnapshots explicitly on a snapshot
This behavior is likely not intuitive for the user, especially in the case where they may want to disable pruning for a dataset, and expect that to inherit to all snapshots of that dataset. Setting it explicitly would require the user to manually call inherit.
The case of wanting a specific snapshot to be set differently than the rest is the outlier, so just inherit this from the parent.