Re-base feature branch on 1.1 branch with squashed commits
Make this more readable
Seal these types

Make setter internal

Not private, so it can be tested

These need to be interlocked

RMW operations are not atomic
Making the increments interlocked is also meaningless if sets are not also interlocked.
Sets on the properties are now disallowed.
The fields are only set by Interlocked.Exchange

Add service start time and pull items up to the interface

Add some endpoints and testing out just using kestrel configuration